+++ src/sample/model/Elevator.java
@@ -1,0 +1,4 @@
+package sample.model;
+
+public class Elevator {
+}
